+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route, BrowserRouter as Router } from 'react-router-dom';
-import { CatalogueComponent } from './features/catalogue/Catalogue';
-import cataloguePage from './features/cataloguePage/cataloguePage';
+import { CatalogueComponent } from './features/catalogue/Component/Catalogue';
+import cataloguePage from './features/cataloguePage/Component/cataloguePage';
 import './App.css';
 
 function App() {
