+++ src/app/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import catalogueReducer from '../features/catalogue/catalogueSlice';
+import catalogueReducer from '../features/catalogue/Reducer/catalogueSlice';
 
 export const store = configureStore({
   reducer: {
